feat(venue): move PracticalInformation in new folder
