+++ src/features/venue/components/PracticalInformation/PracticalInformation.tsx
